+++ trunk/java/src/AZBIrenderer/Sphere.java
@@ -32,7 +32,7 @@ public class Sphere extends SingleMaterialSurface implements Surface {
         R_square = this.radius * this.radius;
         if (d_square > R_square) return false;
 
-        P3P2_length = (double)Math.sqrt(R_square - d_square);
+        P3P2_length = Math.sqrt(R_square - d_square);
 
         intersect.T = P1P2_length - P3P2_length;
 	intersect.point = add(r.origin, mul (intersect.T, r.direction));
@@ -45,8 +45,8 @@ public class Sphere extends SingleMaterialSurface implements Surface {
             double z = InnerProduct(d, Math3D.Zaxis);
             double y = InnerProduct(d, Math3D.Yaxis);
             double x = InnerProduct(d, Math3D.Xaxis);
-            intersect.u = ((double) Math.atan2(y, x) + Math3D.PI) * Math3D.INV_PI2;
-            intersect.v = (double) Math.acos((intersect.point.z - this.center.z) / this.radius) * Math3D.INV_PI;
+            intersect.u = ( Math.atan2(y, x) + Math3D.PI) * Math3D.INV_PI2;
+            intersect.v =  Math.acos((intersect.point.z - this.center.z) / this.radius) * Math3D.INV_PI;
         }
 
         return true;
